Create README.md
I do handcrafted walking sticks, trinket and jewelry boxes, sliced oval wood and square wood wall art with hand drawn designs and motivational qoutes and dedicational events on them. They are then wood burned an colored in to make the design stand out and pop. They are then sealed with a coat of polyurathane to seal and protect them